Testimonial fixed and functions and props moved to context
