+++ frontend/src/components/Testimonials.jsx
@@ -1,113 +1,19 @@
-import React, { useEffect, useState, useContext } from "react";
+import React, { useContext } from "react";
 import { Context } from "../context/Context";
 
 const Testimonials = () => {
-  const { loggedInUserData, userId, isLoggedIn } = useContext(Context);
-  const [testimonials, setTestimonials] = useState([]);
-  const [showModal, setShowModal] = useState(false);
-  const [hasPosted, setHasPosted] = useState(false);
-  const [showAll, setShowAll] = useState(false);
-  const [formData, setFormData] = useState({
-    username: "",
-    rating: 1,
-    comment: "",
-  });
-
-  useEffect(() => {
-    const fetchTestimonials = async () => {
-      try {
-        const response = await fetch(
-          `${import.meta.env.VITE_API}/testimonials`,
-          {
-            method: "GET",
-            headers: {
-              "Content-Type": "application/json",
-            },
-            credentials: "include",
-          }
-        );
-        if (!response.ok) {
-          const errorData = await response.json();
-          throw new Error(errorData.error.message);
-        }
-        const data = await response.json();
-        setTestimonials(data);
-
-        if (userId) {
-          const userTestimonial = data.find(
-            (testimonial) => testimonial.username === loggedInUserData.username
-          );
-          if (userTestimonial) {
-            setHasPosted(true);
-          }
-        }
-      } catch (error) {
-        alert("Error fetching testimonials", error);
-      }
-    };
-
-    fetchTestimonials();
-  }, [userId, loggedInUserData]);
-
-  useEffect(() => {
-    if (!isLoggedIn) {
-      setFormData({ username: "", rating: 1, comment: "" });
-      setHasPosted(false);
-    }
-  }, [isLoggedIn]);
-
-  const handleInputChange = (e) => {
-    const { name, value } = e.target;
-    setFormData({ ...formData, [name]: value });
-  };
-
-  const avatarOptions = [
-    "https://randomuser.me/api/portraits/lego/1.jpg",
-    "https://randomuser.me/api/portraits/lego/2.jpg",
-    "https://randomuser.me/api/portraits/lego/3.jpg",
-    "https://www.pngall.com/wp-content/uploads/12/Avatar-Profile-Vector-PNG-Pic.png",
-    "https://kigali.walk21.com/wp-content/uploads/2023/10/Mesa-de-trabajo-1-e1696881708994.png",
-    "https://cdn0.iconfinder.com/data/icons/user-interface-vol-3-12/66/68-512.png",
-    "https://cdn3.iconfinder.com/data/icons/line-color-mix/66/Mix-22-512.png",
-    "https://cdn0.iconfinder.com/data/icons/seo-and-marketing-volume-1/256/100-512.png",
-    "https://cdn3.iconfinder.com/data/icons/user-interface-2463/24/user_profile_profile_personal_user_person-512.png",
-    "https://cdn0.iconfinder.com/data/icons/user-interface-vol-3-12/66/68-512.png",
-    "https://cdn1.iconfinder.com/data/icons/vibrancie-action/30/action_030-account-user-profile-avatar-256.png",
-  ];
-
-  const getRandomAvatar = () => {
-    const randomIndex = Math.floor(Math.random() * avatarOptions.length);
-    return avatarOptions[randomIndex];
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    const postData = {
-      rating: formData.rating,
-      comment: formData.comment,
-      avatar: userId ? loggedInUserData.profileImage : getRandomAvatar(),
-      username: userId ? loggedInUserData.username : formData.username,
-    };
-
-    try {
-      const response = await fetch(`${import.meta.env.VITE_API}/testimonials`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(postData),
-      });
-      if (!response.ok) {
-        throw new Error("Network response was not ok");
-      }
-      const newTestimonial = await response.json();
-      setTestimonials([...testimonials, newTestimonial]);
-      setShowModal(false);
-      setFormData({ username: "", rating: 1, comment: "" });
-    } catch (error) {
-      console.error("Error submitting testimonial", error);
-    }
-  };
+  const {
+    userId,
+    testimonials,
+    showModal,
+    setShowModal,
+    hasPosted,
+    showAll,
+    setShowAll,
+    testimonialData,
+    handleInputChange,
+    handleSubmit,
+  } = useContext(Context);
 
   return (
     <div className="relative">
@@ -117,6 +23,7 @@ const Testimonials = () => {
         </h2>
         <div className="container mx-auto flex flex-wrap justify-center gap-8">
           {testimonials
+            .toReversed()
             .slice(0, showAll ? testimonials.length : 6)
             .map((testimonial) => (
               <div
@@ -200,7 +107,7 @@ const Testimonials = () => {
                     type="text"
                     id="username"
                     name="username"
-                    value={formData.username}
+                    value={testimonialData.username}
                     onChange={handleInputChange}
                     className="w-full px-3 py-2 border rounded-lg bg-white bg-opacity-30 backdrop-blur-md"
                     required
@@ -214,7 +121,7 @@ const Testimonials = () => {
                 <select
                   id="rating"
                   name="rating"
-                  value={formData.rating}
+                  value={testimonialData.rating}
                   onChange={handleInputChange}
                   className="w-full px-3 py-2 border rounded-lg bg-white bg-opacity-30 backdrop-blur-md"
                   required
@@ -233,7 +140,7 @@ const Testimonials = () => {
                 <textarea
                   id="comment"
                   name="comment"
-                  value={formData.comment}
+                  value={testimonialData.comment}
                   onChange={handleInputChange}
                   className="w-full px-3 py-2 border rounded-lg bg-white bg-opacity-30 backdrop-blur-md"
                   required
